refactor: Move servlet abstraction data structures to data.servletabstraction
